remove gitattributes to prevent Miniver version from corrupting tarballs
People who want to use Miniver will either copy/paste from Github
or download a zip/tar and use that as the skeleton of their project.
In the latter case the `_static_version.py` file would be "corrupted"
with the Miniver version if we had a correct entry in .gitattributes,
and would have to be manually edited. It is preferable to have users
create their own .gitattributes file than have to modify
_static_version.py by hand.